Separate utilities, add constants for sizing
