+++ components/_utilities/size.js
@@ -1,0 +1,11 @@
+// Values for phSize* props:
+const SMALL = 'small';
+const LARGE = 'large';
+
+const isSmall = size => size === SMALL;
+const isLarge = size => size === LARGE;
+
+module.exports = {
+  isSmall,
+  isLarge,
+};
